rematch new game

+++ crates/shared/src/messages/game.rs
@@ -12,15 +12,15 @@ pub enum GameOverReason {
     Resignation,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub enum GameServerMessage {
     UserJoined {
         uuid: Uuid,
         position_fen: String,
         player_role: PlayerRole,
         moves: Vec<String>,
-        white_wins: u32,
-        black_wins: u32,
+        white_wins: f32,
+        black_wins: f32,
     },
     UserLeft {
         username: String,
@@ -55,14 +55,14 @@ pub enum GameServerMessage {
         turn: Side,
         sent_at_ms: i64,
         clock_running: bool,
-        white_wins: u32,
-        black_wins: u32,
+        white_wins: f32,
+        black_wins: f32,
     },
     GameOver {
         winner: Option<Side>,
         reason: GameOverReason,
-        white_wins: u32,
-        black_wins: u32,
+        white_wins: f32,
+        black_wins: f32,
     },
     RematchOffer {
         from: Uuid,
